collect all endpoints into single source file
